% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonparametric-impute.R
\name{bayes_np_impute_groups_safe}
\alias{bayes_np_impute_groups_safe}
\title{Bayesian Nonparametric Imputation for Multiple Groups (Internal)}
\usage{
bayes_np_impute_groups_safe(
  data,
  time_col,
  status_col,
  groups,
  n_imputations,
  prior,
  mcmc,
  grid,
  verbose,
  time_unit = NULL
)
}
\arguments{
\item{data}{A data frame containing survival data}

\item{time_col}{Name of the time column}

\item{status_col}{Name of the status column (1 = event, 0 = censored)}

\item{groups}{Name of the grouping variable}

\item{n_imputations}{Number of imputed complete datasets to generate (default: 10)}

\item{prior}{List of prior parameters for the LDDP model}

\item{mcmc}{List of MCMC sampling options (nburn, nsave, nskip, ndisplay)}

\item{grid}{Prediction grid for survival curves}

\item{verbose}{Print progress messages (default: TRUE)}
}
\value{
A bayesian_imputation_groups object
}
\description{
This function handles group-based nonparametric imputation by fitting separate
LDDP models for each group and combining the results.
}
\keyword{internal}
