% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonparametric-impute.R
\name{bayes_np_impute_single}
\alias{bayes_np_impute_single}
\title{Bayesian Nonparametric Imputation for Single Group (Internal)}
\usage{
bayes_np_impute_single(
  data,
  time_col = "time",
  status_col = "status",
  n_imputations = 10,
  prior = NULL,
  mcmc = list(nburn = 1000, nsave = 4000, nskip = 1, ndisplay = 100),
  grid = NULL,
  verbose = TRUE,
  time_unit = NULL
)
}
\arguments{
\item{data}{A data frame containing survival data}

\item{time_col}{Name of the time column}

\item{status_col}{Name of the status column (1 = event, 0 = censored)}

\item{n_imputations}{Number of imputed complete datasets to generate (default: 10)}

\item{prior}{List of prior parameters for the LDDP model}

\item{mcmc}{List of MCMC sampling options (nburn, nsave, nskip, ndisplay)}

\item{grid}{Prediction grid for survival curves}

\item{verbose}{Print progress messages (default: TRUE)}
}
\value{
A bayesian_imputation object
}
\description{
Original bayes_np_impute logic, now separated for single-group analysis.
This function is called internally by bayes_np_impute() when no groups are specified.
}
\keyword{internal}
