% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian-impute.R
\name{bayesian_impute_groups_safe}
\alias{bayesian_impute_groups_safe}
\title{Safe Group Processing Function}
\usage{
bayesian_impute_groups_safe(
  data,
  time_col,
  status_col,
  groups,
  n_imputations,
  distribution,
  priors,
  mcmc_options,
  verbose,
  time_unit = NULL
)
}
\arguments{
\item{data}{Data frame containing survival data}

\item{time_col}{Name of the time column}

\item{status_col}{Name of the status column}

\item{groups}{Name of the group variable}

\item{n_imputations}{Number of imputations}

\item{distribution}{Distribution to use}

\item{priors}{Prior parameters}

\item{mcmc_options}{MCMC options}

\item{verbose}{Print progress messages}
}
\value{
Group comparison object
}
\description{
Processes groups individually with error isolation
}
\keyword{internal}
