% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian-impute.R
\name{bayesian_impute_single}
\alias{bayesian_impute_single}
\title{Bayesian Imputation for Single Group (Internal)}
\usage{
bayesian_impute_single(
  data,
  time_col = NULL,
  status_col = NULL,
  n_imputations = 10,
  distribution = "weibull",
  priors = NULL,
  mcmc_options = NULL,
  verbose = TRUE,
  time_unit = NULL
)
}
\arguments{
\item{data}{A data frame containing survival data}

\item{time_col}{Name of the time column}

\item{status_col}{Name of the status column (1 = event, 0 = censored)}

\item{n_imputations}{Number of imputed complete datasets to generate (default: 10)}

\item{distribution}{Distribution to use ("weibull", "exponential", "lognormal")}

\item{priors}{List of prior parameters for the distribution parameters}

\item{mcmc_options}{List of MCMC sampling options (chains, iterations, etc.)}

\item{verbose}{Print progress messages (default: TRUE)}
}
\value{
A bayesian_imputation object
}
\description{
This is the bayesian_impute logic, separated for single-group analysis.
This function is called internally by bayesian_impute() when no groups are specified.
}
\keyword{internal}
