% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical-metrics.R
\name{calculate_clinical_metrics}
\alias{calculate_clinical_metrics}
\title{Calculate survival statistics for original and imputed data}
\usage{
calculate_clinical_metrics(x, time_points = c(365, 730, 1825))
}
\arguments{
\item{x}{bayesian_imputation object}

\item{time_points}{Vector of time points for survival probability estimates}
}
\value{
List with clinical metrics
}
\description{
Calculate survival statistics for original and imputed data
}
\keyword{internal}
