% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical-metrics.R
\name{calculate_imputation_summary}
\alias{calculate_imputation_summary}
\title{Calculate Clinical Survival Metrics}
\usage{
calculate_imputation_summary(x)
}
\arguments{
\item{x}{bayesian_imputation object}
}
\value{
List with imputation metrics
}
\description{
Helper functions to compute clinically meaningful survival statistics
from bayesian_imputation objects
}
\keyword{Calculate}
\keyword{censored}
\keyword{comparing}
\keyword{imputation}
\keyword{imputed}
\keyword{internal}
\keyword{summary}
\keyword{values}
\keyword{vs}
