% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{calculate_log_rank_test}
\alias{calculate_log_rank_test}
\title{Calculate pooled Cox Wald test (log-rank equivalent) for group comparison}
\usage{
calculate_log_rank_test(object)
}
\arguments{
\item{object}{bayesian_imputation_groups object}
}
\description{
Calculate pooled Cox Wald test (log-rank equivalent) for group comparison
}
\keyword{internal}
