% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{calculate_median_survival_diff}
\alias{calculate_median_survival_diff}
\title{Calculate median survival difference between groups}
\usage{
calculate_median_survival_diff(object)
}
\arguments{
\item{object}{bayesian_imputation_groups object}
}
\description{
Calculate median survival difference between groups
}
\keyword{internal}
