% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{combine_group_datasets_safe}
\alias{combine_group_datasets_safe}
\title{Combine group datasets safely}
\usage{
combine_group_datasets_safe(object, dataset, format)
}
\arguments{
\item{object}{bayesian_imputation_groups object}

\item{dataset}{Dataset number}

\item{format}{Output format}
}
\description{
Combine group datasets safely
}
\keyword{internal}
