% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete.R
\name{complete}
\alias{complete}
\title{Extract completed datasets}
\usage{
complete(object, ...)
}
\arguments{
\item{object}{Object produced by \code{\link[=bayesian_impute]{bayesian_impute()}} or related helpers.}

\item{...}{Passed to methods.}
}
\description{
Generic function to extract completed datasets from a
\code{bayesian_imputation} or \code{bayesian_imputation_groups} object. See the
method documentation for details.
}
