% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save-imputed-data.R
\name{export_group_imputed_data}
\alias{export_group_imputed_data}
\title{Export Group Imputed Data}
\usage{
export_group_imputed_data(
  object,
  file_path,
  format,
  datasets,
  groups,
  include_original,
  combine,
  ...
)
}
\description{
Export Group Imputed Data
}
\keyword{internal}
