% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_time}
\alias{format_time}
\title{Format time for display}
\usage{
format_time(time_seconds)
}
\arguments{
\item{time_seconds}{Time in seconds}
}
\value{
Formatted time string
}
\description{
Format time for display
}
\keyword{internal}
