% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_distribution_param_labels}
\alias{get_distribution_param_labels}
\title{Get parameter labels for plotting}
\usage{
get_distribution_param_labels(distribution)
}
\arguments{
\item{distribution}{Distribution name ("weibull", "exponential", "lognormal")}
}
\value{
Character vector of parameter labels for plots
}
\description{
Get parameter labels for plotting
}
\keyword{internal}
