% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_distribution_params}
\alias{get_distribution_params}
\title{Get parameter names for a distribution}
\usage{
get_distribution_params(distribution)
}
\arguments{
\item{distribution}{Distribution name ("weibull", "exponential", "lognormal")}
}
\value{
Character vector of parameter names
}
\description{
Get parameter names for a distribution
}
\keyword{internal}
