% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-models.R
\name{get_stan_model}
\alias{get_stan_model}
\title{Get compiled Stan model}
\usage{
get_stan_model(model_name)
}
\arguments{
\item{model_name}{Name of the model (e.g., "weibull")}
}
\value{
Compiled Stan model object
}
\description{
Get compiled Stan model
}
\keyword{internal}
