% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.bayesian_imputation}
\alias{plot.bayesian_imputation}
\title{Plot method for bayesian_imputation objects}
\usage{
\method{plot}{bayesian_imputation}(x, type = "survival", ...)
}
\arguments{
\item{x}{A bayesian_imputation object}

\item{type}{Type of plot: "survival", "trace", "pairs", "posterior", "completed_dataset_summary", "boxplots_comparison"}

\item{...}{Additional arguments passed to plotting functions}
}
\description{
Plot method for bayesian_imputation objects
}
