% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDDPsurvival.R
\name{plot.LDDPsurvival}
\alias{plot.LDDPsurvival}
\title{Plot method for LDDPsurvival objects}
\usage{
\method{plot}{LDDPsurvival}(
  x,
  hpd = TRUE,
  ask = TRUE,
  nfigr = 2,
  nfigc = 2,
  param = NULL,
  col = "#bdfcc9",
  ...
)
}
\arguments{
\item{x}{LDDPsurvival object}

\item{hpd}{logical, whether to show HPD intervals}

\item{ask}{logical, whether to ask before plotting}

\item{nfigr}{number of figure rows}

\item{nfigc}{number of figure columns}

\item{param}{parameter to plot}

\item{col}{color for plots}

\item{...}{additional arguments}
}
\description{
Plot method for LDDPsurvival objects
}
