% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-helpers.R
\name{plot_boxplots_comparison}
\alias{plot_boxplots_comparison}
\title{Plot boxplots comparison of all completed datasets}
\usage{
plot_boxplots_comparison(x, n_max = 10, dataset_indices = NULL, ...)
}
\arguments{
\item{x}{bayesian_imputation object}

\item{n_max}{Number of datasets to sample for boxplots (default: 10)}

\item{dataset_indices}{Optional indices of datasets to include (random if NULL)}
}
\value{
ggplot object
}
\description{
Plot boxplots comparison of all completed datasets
}
\keyword{internal}
