% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-helpers.R
\name{plot_density_comparison}
\alias{plot_density_comparison}
\title{Plot density comparison using logspline}
\usage{
plot_density_comparison(x, n_curves = NULL, alpha = 0.3, ...)
}
\arguments{
\item{x}{bayesian_imputation object}

\item{n_curves}{Number of imputed curves to show (NULL for all)}

\item{alpha}{Transparency for imputed curves}

\item{...}{Additional arguments}
}
\value{
ggplot object
}
\description{
Plot density comparison using logspline
}
\keyword{internal}
