% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-helpers.R
\name{plot_group_comparison_panels_safe}
\alias{plot_group_comparison_panels_safe}
\title{Plot group comparison panels}
\usage{
plot_group_comparison_panels_safe(x, ...)
}
\arguments{
\item{x}{bayesian_imputation_groups object}

\item{...}{Additional arguments}
}
\description{
Plot group comparison panels
}
\keyword{internal}
