% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-helpers.R
\name{plot_individual_groups_safe}
\alias{plot_individual_groups_safe}
\title{Plot individual groups}
\usage{
plot_individual_groups_safe(x, type, ...)
}
\arguments{
\item{x}{bayesian_imputation_groups object}

\item{type}{Plot type}

\item{...}{Additional arguments}
}
\description{
Plot individual groups
}
\keyword{internal}
