% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-helpers.R
\name{plot_pairs}
\alias{plot_pairs}
\title{Plot parameter correlation pairs}
\usage{
plot_pairs(x, ...)
}
\arguments{
\item{x}{bayesian_imputation object}
}
\value{
ggplot object or NULL
}
\description{
Plot parameter correlation pairs
}
\keyword{internal}
