% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-helpers.R
\name{plot_trace_plots}
\alias{plot_trace_plots}
\title{Plot MCMC trace plots for model parameters}
\usage{
plot_trace_plots(x, ...)
}
\arguments{
\item{x}{bayesian_imputation object}
}
\value{
ggplot object or NULL
}
\description{
Plot MCMC trace plots for model parameters
}
\keyword{internal}
