% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{pool_cox_group_test}
\alias{pool_cox_group_test}
\title{Pool Cox PH group comparison across imputations (Rubin's rules)}
\usage{
pool_cox_group_test(object)
}
\arguments{
\item{object}{bayesian_imputation_groups object}
}
\description{
Pool Cox PH group comparison across imputations (Rubin's rules)
}
\keyword{internal}
