% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.bayesian_imputation}
\alias{print.bayesian_imputation}
\title{Print method for bayesian_imputation objects}
\usage{
\method{print}{bayesian_imputation}(x, ...)
}
\arguments{
\item{x}{A bayesian_imputation object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for bayesian_imputation objects
}
