% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.bayesian_imputation_groups}
\alias{print.bayesian_imputation_groups}
\title{Print method for bayesian_imputation_groups objects}
\usage{
\method{print}{bayesian_imputation_groups}(x, ...)
}
\arguments{
\item{x}{A bayesian_imputation_groups object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for bayesian_imputation_groups objects
}
