% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDDPsurvival.R
\name{print.LDDPsurvival}
\alias{print.LDDPsurvival}
\title{Print method for LDDPsurvival objects}
\usage{
\method{print}{LDDPsurvival}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{LDDPsurvival object}

\item{digits}{number of digits to print}

\item{...}{additional arguments}
}
\description{
Print method for LDDPsurvival objects
}
