% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDDPsurvival.R
\name{print.summaryLDDPsurvival}
\alias{print.summaryLDDPsurvival}
\title{Print method for summaryLDDPsurvival objects}
\usage{
\method{print}{summaryLDDPsurvival}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{summaryLDDPsurvival object}

\item{digits}{number of digits to print}

\item{...}{additional arguments}
}
\description{
Print method for summaryLDDPsurvival objects
}
