% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print_comprehensive_group_comparison_safe}
\alias{print_comprehensive_group_comparison_safe}
\title{Print comprehensive group comparison (with statistical tests)}
\usage{
print_comprehensive_group_comparison_safe(object)
}
\arguments{
\item{object}{bayesian_imputation_groups object}
}
\description{
Print comprehensive group comparison (with statistical tests)
}
\keyword{internal}
