% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print_group_overview_safe}
\alias{print_group_overview_safe}
\title{Print group overview with key metrics}
\usage{
print_group_overview_safe(x)
}
\arguments{
\item{x}{bayesian_imputation_groups object}
}
\description{
Print group overview with key metrics
}
\keyword{internal}
