% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print_survival_comparison_safe}
\alias{print_survival_comparison_safe}
\title{Print survival comparison between groups}
\usage{
print_survival_comparison_safe(x, group_metrics)
}
\arguments{
\item{x}{bayesian_imputation_groups object}

\item{group_metrics}{List of clinical metrics for each group}
}
\description{
Print survival comparison between groups
}
\keyword{internal}
