% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save-imputed-data.R
\name{remove_original_columns}
\alias{remove_original_columns}
\title{Remove Original Value Columns}
\usage{
remove_original_columns(data)
}
\description{
Remove Original Value Columns
}
\keyword{internal}
