% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-models.R
\name{stan_models_ready}
\alias{stan_models_ready}
\title{Check if Stan models are compiled}
\usage{
stan_models_ready()
}
\value{
Logical indicating if models are ready
}
\description{
Check if Stan models are compiled
}
\keyword{internal}
