% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDDPsurvival.R
\name{summary.LDDPsurvival}
\alias{summary.LDDPsurvival}
\title{Summary method for LDDPsurvival objects}
\usage{
\method{summary}{LDDPsurvival}(object, hpd = TRUE, ...)
}
\arguments{
\item{object}{LDDPsurvival object}

\item{hpd}{logical, whether to compute HPD intervals}

\item{...}{additional arguments}
}
\description{
Summary method for LDDPsurvival objects
}
