% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian-impute.R
\name{validate_groups_safe}
\alias{validate_groups_safe}
\title{Safe Group Validation}
\usage{
validate_groups_safe(data, groups, time_col, status_col, verbose = TRUE)
}
\arguments{
\item{data}{Data frame containing survival data}

\item{groups}{Name of the group variable}

\item{time_col}{Name of the time column}

\item{status_col}{Name of the status column}

\item{verbose}{Print progress messages}
}
\value{
List with validation results
}
\description{
Validates group variables and data structure before processing
}
\keyword{internal}
