% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-validation.R
\name{validate_imputation_params}
\alias{validate_imputation_params}
\title{Validate imputation parameters}
\usage{
validate_imputation_params(n_imputations, distribution, mcmc_options)
}
\arguments{
\item{n_imputations}{Number of imputations}

\item{distribution}{Distribution name}

\item{mcmc_options}{MCMC options list}
}
\value{
Validated parameters list
}
\description{
Validate imputation parameters
}
\keyword{internal}
